Separate out Moldfile into file module
